Fixed Backend Issues

+++ backend/src/main/java/com/devdash/backend/service/JobService.java
@@ -52,7 +52,7 @@ public class JobService {
                 .worker(worker)
                 .scheduledStart(dto.getScheduledStart())
                 .scheduledEnd(dto.getScheduledEnd())
-                .finalPrice(dto.getEstimatedPrice())
+                .finalPrice(dto.getEstimatedPrice().doubleValue())
                 .build();
 
         assignment = assignmentRepo.save(assignment);
